src/requirements.txt
cloudshell-shell-core>=4.0.0,<4.1.0
cloudshell-cp-core>=1.0.0,<1.1.0
cloudshell-automation-api>=8.3.0.0,<8.3.1.0
requests